(issue 66) used long instead of "long long"

+++ psutil/_psutil_bsd.c
@@ -330,7 +330,7 @@ static PyObject* get_memory_info(PyObject* self, PyObject* args)
 static PyObject* get_total_phymem(PyObject* self, PyObject* args)
 {
     #ifdef __x86_64__
-        long long total_phymem;
+        long total_phymem;
     #else
         int total_phymem;
     #endif
